+++ onmt/modules/Decoder.lua
@@ -310,21 +310,21 @@ end
   Parameters:
 
   * `batch` - `Batch` object
-  * `encoderStates` -
+  * `initialStates` - initialization of decoder.
   * `context` -
   * `func` - Calls `func(out, t)` each timestep.
 --]]
 
-function Decoder:forwardAndApply(batch, encoderStates, context, func)
+function Decoder:forwardAndApply(batch, initialStates, context, func)
   -- TODO: Make this a private method.
 
   if self.statesProto == nil then
-    self.statesProto = onmt.utils.Tensor.initTensorTable(#encoderStates,
+    self.statesProto = onmt.utils.Tensor.initTensorTable(#initialStates,
                                                          self.stateProto,
                                                          { batch.size, self.args.rnnSize })
   end
 
-  local states = onmt.utils.Tensor.copyTensorTable(self.statesProto, encoderStates)
+  local states = onmt.utils.Tensor.copyTensorTable(self.statesProto, initialStates)
 
   local prevOut
 
@@ -339,13 +339,13 @@ end
   Parameters:
 
   * `batch` - a `Batch` object.
-  * `encoderStates` - a batch of initial decoder states (optional) [0]
+  * `initialStates` - a batch of initial decoder states (optional) [0]
   * `context` - the context to apply attention to.
 
   Returns: Table of top hidden state for each timestep.
 --]]
-function Decoder:forward(batch, encoderStates, context)
-  encoderStates = encoderStates
+function Decoder:forward(batch, initialStates, context)
+  initialStates = initialStates
     or onmt.utils.Tensor.initTensorTable(self.args.numEffectiveLayers,
                                          onmt.utils.Cuda.convert(torch.Tensor()),
                                          { batch.size, self.args.rnnSize })
@@ -355,7 +355,7 @@ function Decoder:forward(batch, encoderStates, context)
 
   local outputs = {}
 
-  self:forwardAndApply(batch, encoderStates, context, function (out)
+  self:forwardAndApply(batch, initialStates, context, function (out)
     table.insert(outputs, out)
   end)
 
@@ -452,19 +452,19 @@ end
 Parameters:
 
   * `batch` - a `Batch` to score.
-  * `encoderStates` - initialization of decoder.
+  * `initialStates` - initialization of decoder.
   * `context` - the attention context.
   * `criterion` - a pointwise criterion.
 
 --]]
-function Decoder:computeLoss(batch, encoderStates, context, criterion)
-  encoderStates = encoderStates
+function Decoder:computeLoss(batch, initialStates, context, criterion)
+  initialStates = initialStates
     or onmt.utils.Tensor.initTensorTable(self.args.numEffectiveLayers,
                                          onmt.utils.Cuda.convert(torch.Tensor()),
                                          { batch.size, self.args.rnnSize })
 
   local loss = 0
-  self:forwardAndApply(batch, encoderStates, context, function (out, t)
+  self:forwardAndApply(batch, initialStates, context, function (out, t)
     local pred = self.generator:forward(out)
     local output = batch:getTargetOutput(t)
     loss = loss + criterion:forward(pred, output)
@@ -479,19 +479,19 @@ end
 Parameters:
 
   * `batch` - a `Batch` to score.
-  * `encoderStates` - initialization of decoder.
+  * `initialStates` - initialization of decoder.
   * `context` - the attention context.
 
 --]]
-function Decoder:computeScore(batch, encoderStates, context)
-  encoderStates = encoderStates
+function Decoder:computeScore(batch, initialStates, context)
+  initialStates = initialStates
     or onmt.utils.Tensor.initTensorTable(self.args.numEffectiveLayers,
                                          onmt.utils.Cuda.convert(torch.Tensor()),
                                          { batch.size, self.args.rnnSize })
 
   local score = {}
 
-  self:forwardAndApply(batch, encoderStates, context, function (out, t)
+  self:forwardAndApply(batch, initialStates, context, function (out, t)
     local pred = self.generator:forward(out)
     for b = 1, batch.size do
       if t <= batch.targetSize[b] then